Refactor whole app
The idea is to create a nice `Graph` that one can easily work with for
analysis, instead of doing that directly on react flow data, which is
annoying and slow. Now the react flow data is created from the internal
graph on each render instead.

Generally the whole code base is a bit nicer now, as "business logic"
code is mainly in `graph/` now, not scattered around the code base.

This is not just a refactor though, a few things changed as well. For
example, the IDs for the graph are now numbers, and the string IDs for
react-flow changed a bit. Further, you cannot add more than one edge
per handle.

This is not yet done, two things are still missing:
- Properly implement `isValidConnection`
- Bring back persistance