Re-add persistance with localStorage
So this requires quite a bit of new code. Combined with the previous
commits, the code base grew quite a bit to have essentially the same
functionality. So was it worth it? I still think yes. Analysis code on
the graph is easier to write now. And even though we split the previous
state into three "levels" (our actual state, the react-flow data we
generate from it on the flow, and the persistance data), I think that's
nicer. Our actual state is just precisely what we want as state, not
containing any random other data (like node selection data before). I
also like having the storage format be explicitly defined and decoupled
from internal data structures.